[SWM-175] Move : ui components
