+++ src/components/search/SearchLectureCard.tsx
